Change license to MIT (from GPL3)
After realizing that GPL3 license doesn't give permission to use the files commercially, it was changed to MIT license for more freedom to users although no one will ever see this repository. Also, MIT license needs you to put the license file in your copy of this code, don't need to put that.